+++ build.sh
@@ -24,7 +24,7 @@ for ARCH in arm64 x86_64; do
 		Sources/Views.swift \
 		Sources/BgviewerApp.swift \
 		-framework SwiftUI -framework AppKit -framework Foundation \
-		-framework ServiceManagement \
+		-framework ServiceManagement -framework UserNotifications \
 		-o "$BIN-$ARCH"
 	SLICES+=("$BIN-$ARCH")
 done